backend/requirements.txt
# FastAPI
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
pydantic==2.10.4
pydantic[email]==2.10.4
pydantic-settings==2.8.1
email-validator==2.2.0

# Firebase
firebase-admin==6.6.0

# Autenticação
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Utilitários
python-dateutil==2.9.0.post0
Pillow==11.1.0

# Development
pytest==8.3.4
pytest-asyncio==0.25.0